Tweak "edit on GitHub" button
The button seems to have regressed in recent releases of the Hotdoc
theme. It's too large and its capitalization looks weird. Fix these.

Additionally, make it an accent color instead of the primary brand color.
That makes it blend in better with the right-hand navigation bar, it
shouldn't be so prominent.

https://phabricator.endlessm.com/T21126